requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.0
djangorestframework==3.15.1
filelock==3.13.1
platformdirs==4.2.0
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.25.1
